requirements.txt
astroid==2.3.3
bcrypt==3.1.7
cffi==1.13.2
Click==7.0
Flask==1.1.1
Flask-PyMongo==2.3.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.4
pycparser==2.19
pylint==2.4.3
pymongo==3.9.0
python-dotenv==0.10.3
six==1.13.0
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2
